# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_ayMissouri_watchlist-go_git_internal_models.ClientEvent:
    properties:
      episode:
        type: integer
      event_type:
        type: string
      genres:
        items:
          type: string
        type: array
      imdb_id:
        type: string
      item_id:
        type: string
      media_type:
        type: string
      metadata:
        additionalProperties: {}
        type: object
      occurred_at:
        type: integer
      release_year:
        type: integer
      runtime_minutes:
        type: integer
      season:
        type: integer
      title:
        type: string
    type: object
...
  github_com_ayMissouri_watchlist-go_git_internal_models.EventsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.UserEvent'
        type: array
    type: object
...
  github_com_ayMissouri_watchlist-go_git_internal_models.RecordEventsRequest:
    properties:
      events:
        items:
          $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.ClientEvent'
        type: array
    type: object
...
  github_com_ayMissouri_watchlist-go_git_internal_models.UserEvent:
    properties:
      episode:
        type: integer
      event_type:
        type: string
      genres:
        items:
          type: string
        type: array
      id:
        type: integer
      imdb_id:
        type: string
      item_id:
        type: string
      media_type:
        type: string
      metadata:
        additionalProperties: {}
        type: object
      occurred_at:
        type: integer
      release_year:
        type: integer
      runtime_minutes:
        type: integer
      season:
        type: integer
      source:
        type: string
      title:
        type: string
      title_english:
        description: read-only, from the anime's watchlist item
        type: string
      user_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: ayMissouri
    url: https://github.com/ayMissouri/watchlist-go.git
  description: Backend API for tracking movies and TV shows.
  license:
    name: MIT
  title: Watchlist API
  version: "1.0"
paths:
...
  /events:
    get:
      description: The user's latest events, newest first. Meant for an activity feed,
        or just for showing people what we've recorded about them.
      parameters:
      - default: 50
        description: Max events to return
        in: query
        name: limit
        type: integer
      - description: Comma-separated media types to keep (tv, movie, anime)
        in: query
        name: media_type
        type: string
      - description: Only include events at or after this Unix ms timestamp
        in: query
        name: from
        type: integer
      - description: Only include events strictly before this Unix ms timestamp
        in: query
        name: to
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.EventsResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Recent activity
      tags:
      - events
    post:
      consumes:
      - application/json
      description: Takes a batch of events from the client (play/pause, heartbeats,
        page views, that kind of thing) and appends them to the user's activity log.
        Anything with a blank or absurdly long type is dropped quietly.
      parameters:
      - description: Batch of events
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.RecordEventsRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Ingest activity events
      tags:
      - events
...
securityDefinitions:
  BearerAuth:
    description: 'Enter your JWT token as: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"