# internal/handlers/events.go
package handlers

import (
	"encoding/json"
	"net/http"
	"strconv"

	"github.com/ayMissouri/watchlist-go.git/internal/db"
	"github.com/ayMissouri/watchlist-go.git/internal/middleware"
	"github.com/ayMissouri/watchlist-go.git/internal/models"
	"github.com/ayMissouri/watchlist-go.git/internal/tracking"
)

const (
	defaultEventsLimit = 50
	maxEventsLimit     = 200
)

type EventsHandler struct {
	DB      *db.DB
	Tracker *tracking.Service
}

// Record godoc
// @Summary     Ingest activity events
// @Description Takes a batch of events from the client (play/pause, heartbeats, page views, that kind of thing) and appends them to the user's activity log. Anything with a blank or absurdly long type is dropped quietly.
// @Tags        events
// @Accept      json
// @Param       body body models.RecordEventsRequest true "Batch of events"
// @Success     204
// @Failure     400 {object} map[string]string
// @Failure     401 {object} map[string]string
// @Security    BearerAuth
// @Router      /events [post]
func (h *EventsHandler) Record(w http.ResponseWriter, r *http.Request) {
	claims := middleware.ClaimsFromCtx(r)

	var req models.RecordEventsRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		jsonError(w, "invalid request body", http.StatusBadRequest)
		return
	}

	events, ok := tracking.ValidClientEvents(req.Events)
	if !ok {
		jsonError(w, "events must be a non-empty batch of at most 100 items", http.StatusBadRequest)
		return
	}

	h.Tracker.RecordClientEvents(r.Context(), claims.UserID, events)
	w.WriteHeader(http.StatusNoContent)
}

// List godoc
// @Summary     Recent activity
// @Description The user's latest events, newest first. Meant for an activity feed, or just for showing people what we've recorded about them.
// @Tags        events
// @Produce     json
// @Param       limit query int false "Max events to return" default(50)
// @Param       media_type query string false "Comma-separated media types to keep (tv, movie, anime)"
// @Param       from  query int false "Only include events at or after this Unix ms timestamp"
// @Param       to    query int false "Only include events strictly before this Unix ms timestamp"
// @Success     200 {object} models.EventsResponse
// @Failure     401 {object} map[string]string
// @Security    BearerAuth
// @Router      /events [get]
func (h *EventsHandler) List(w http.ResponseWriter, r *http.Request) {
	claims := middleware.ClaimsFromCtx(r)

	limit := defaultEventsLimit
	if l := r.URL.Query().Get("limit"); l != "" {
		if v, err := strconv.Atoi(l); err == nil && v > 0 {
			if v > maxEventsLimit {
				v = maxEventsLimit
			}
			limit = v
		}
	}

	from := parseEpochMs(r.URL.Query().Get("from"))
	to := parseEpochMs(r.URL.Query().Get("to"))

	mediaTypes := parseMediaTypes(r.URL.Query().Get("media_type"))

	items, err := h.DB.GetEvents(r.Context(), claims.UserID, limit, from, to, mediaTypes)
	if err != nil {
		jsonError(w, "could not fetch events", http.StatusInternalServerError)
		return
	}

	jsonOK(w, models.EventsResponse{Items: items})
}

func parseEpochMs(s string) int64 {
	if s == "" {
		return 0
	}
	v, err := strconv.ParseInt(s, 10, 64)
	if err != nil || v < 0 {
		return 0
	}
	return v
}